@@ -1,5 +1,5 @@
 name: digia_ui
-description: Digia UI Engine.
+description: Flutter SDK for rendering server-driven UIs using JSON from Digia Studio. Supports dynamic layouts, state, actions, and custom widgets.
 repository: https://github.com/Digia-Technology-Private-Limited/digia_ui.git
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -19,7 +19,7 @@ repository: https://github.com/Digia-Technology-Private-Limited/digia_ui.git
 version: 1.0.0-rc.9
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: '>=3.3.0 <4.0.0'
 
 # https://github.com/frencojobs/derry
 scripts:
@@ -44,64 +44,52 @@ scripts:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  cached_network_image: ^3.3.0
-  carousel_slider: ^5.0.0
+  cached_network_image: 3.3.1
+  carousel_slider: 5.1.1
   smooth_page_indicator: ^1.2.0+3
-  chewie: 1.8.5
+  chewie: ^1.10.0
   collection: ^1.17.1
-  copy_with_extension: ^6.0.0
-  copy_with_extension_gen: ^6.0.0
-  cupertino_icons: ^1.0.2
-  derry: ^1.5.0
   dio: ^5.3.3
-  dio_intercept_to_curl: ^0.2.0
   expandable: ^5.0.1
   flutter:
     sdk: flutter
   path_drawing: ^1.0.0
-  flutter_blurhash: ^0.8.2
-  flutter_cache_manager: ^3.3.0
+  flutter_blurhash: ^0.9.1
   flutter_html: ^3.0.0-beta.2
   digia_expr: 0.0.14
-  flutter_js: 0.8.2
+  flutter_js: 0.8.5
   fluttertoast: ^8.2.5
-  font_awesome_flutter: ^10.6.0
-  intl: ^0.19.0
+  font_awesome_flutter: 10.8.0
   json_annotation: ^4.9.0
-  lottie: ^3.0.0
+  lottie: 3.0.0
   octo_image: ^2.0.0
   package_info_plus: ^8.0.0
-  flutter_svg: ^2.0.17
+  flutter_svg: 2.1.0
   path_provider: ^2.0.14
-  shared_preferences: ^2.2.2
-  styled_divider: ^1.0.5
-  # timelines: ^0.1.0
+  shared_preferences: ^2.5.3
   url_launcher: ^6.1.11
   uuid: ^4.4.0
   video_player: ^2.8.6
-  webview_flutter: ^4.7.0
+  webview_flutter: 4.7.0
   youtube_player_iframe: ^5.1.1
   table_calendar: ^3.0.9
   pinput: ^5.0.0
   smart_auth: ^2.0.0
   infinite_scroll_pagination: ^4.0.0
-  share_plus: ^10.1.3
-  fl_chart: ^0.69.2
+  share_plus: ^11.0.0
   widget_marquee: ^0.0.8
   before_after: ^3.1.0
   imageview360: ^1.3.0
-  mime_type: ^1.0.0
+  mime_type: ^1.0.1
   http_parser: ^4.0.2
-  file_picker: ^8.0.5
-  talker: ^4.3.4
-  percent_indicator: ^4.2.3
+  file_picker: ^10.2.1
+  percent_indicator: ^4.2.5
   cross_file: ^0.3.4+2
   image_picker: ^1.1.2
-  mocktail: ^1.0.4
   flutter_avif: ^3.0.0
   markdown_widget: ^2.3.2+8
   flutter_staggered_grid_view: ^0.7.0
-  # flutter_math_fork: ^0.7.3
+  web: ^1.1.1
 
 dev_dependencies:
   build_runner: ^2.4.4
@@ -110,6 +98,8 @@ dev_dependencies:
     sdk: flutter
   json_serializable: ^6.7.0
   flutter_lints: ^5.0.0
+  mocktail: ^1.0.4
+  derry: ^1.5.0
 
 # dependency_overrides:
 #   digia_expr:
