@@ -1,5 +1,6 @@
 import 'package:flutter/widgets.dart';
 
+import '../analytics/dui_analytics.dart';
 import 'message_bus.dart';
 
 /// Provides access to Digia UI SDK resources through the widget tree.
@@ -17,13 +18,12 @@ import 'message_bus.dart';
 ///   ),
 /// )
 /// ```
-class DigiaUIScope extends StatefulWidget {
+class DigiaUIScope extends InheritedWidget {
   /// The message bus instance used for communication within the SDK.
   /// If not provided, a new [MessageBus] instance will be created.
   final MessageBus messageBus;
 
-  /// The widget below this widget in the tree.
-  final Widget child;
+  final DUIAnalytics? analyticsHandler;
 
   /// Creates a [DigiaUIScope] widget.
   ///
@@ -32,16 +32,17 @@ class DigiaUIScope extends StatefulWidget {
   DigiaUIScope({
     super.key,
     MessageBus? messageBus,
-    required this.child,
+    this.analyticsHandler,
+    required super.child,
   }) : messageBus = messageBus ?? MessageBus();
 
-  /// Provides access to the current [_DigiaUIScopeInherited] instance.
+  /// Provides access to the current [DigiaUIScope] instance.
   ///
   /// This method is used to access SDK features from descendant widgets.
   /// Will throw an assertion error if no [DigiaUIScope] is found in the widget tree.
-  static _DigiaUIScopeInherited of(BuildContext context) {
+  static DigiaUIScope of(BuildContext context) {
     final inherited =
-        context.dependOnInheritedWidgetOfExactType<_DigiaUIScopeInherited>();
+        context.dependOnInheritedWidgetOfExactType<DigiaUIScope>();
     assert(inherited != null, '''
 No DigiaUIScope found in context. 
 Wrap your app with DigiaUIScope to access SDK features''');
@@ -49,48 +50,7 @@ Wrap your app with DigiaUIScope to access SDK features''');
   }
 
   @override
-  State<DigiaUIScope> createState() => _DigiaUIScopeState();
-}
-
-/// The state for [DigiaUIScope].
-class _DigiaUIScopeState extends State<DigiaUIScope> {
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    widget.messageBus.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return _DigiaUIScopeInherited(
-      messageBus: widget.messageBus,
-      child: widget.child,
-    );
-  }
-}
-
-/// An [InheritedWidget] that provides access to the [MessageBus] instance.
-///
-/// This class is used internally by [DigiaUIScope] to propagate the message bus
-/// down the widget tree. You typically don't need to use this class directly.
-class _DigiaUIScopeInherited extends InheritedWidget {
-  /// The message bus instance that will be made available to descendants.
-  final MessageBus messageBus;
-
-  /// Creates a [_DigiaUIScopeInherited] widget.
-  ///
-  /// Both [messageBus] and [child] parameters must not be null.
-  const _DigiaUIScopeInherited({
-    required this.messageBus,
-    required super.child,
-  });
-
-  @override
-  bool updateShouldNotify(_DigiaUIScopeInherited oldWidget) =>
-      messageBus != oldWidget.messageBus;
+  bool updateShouldNotify(DigiaUIScope oldWidget) =>
+      messageBus != oldWidget.messageBus ||
+      analyticsHandler != oldWidget.analyticsHandler;
 }
