@@ -81,7 +81,7 @@ Future<T?> presentBottomSheet<T>({
                             width: 24,
                             decoration: BoxDecoration(
                                 borderRadius: BorderRadius.circular(12),
-                                color: Colors.white.withOpacity(0.1)),
+                                color: Colors.white.withValues(alpha: 0.1)),
                             child: iconBuilder(innerContext),
                           ),
                         ),
