@@ -1,51 +1,164 @@
 import 'package:digia_ui/digia_ui.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-
-// const String baseUrl = 'http://localhost:3000/api/v1';
-const String baseUrl = 'https://dev.digia.tech/api/v1';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  runApp(DUIApp(
-      digiaAccessKey: "681b8775af81b5e02c8d9f0a",
-      baseUrl: baseUrl,
-      flavorInfo: Debug('main'),
-      // environmentInfo: Versioned(19),
-      // environmentInfo: Production(PrioritizeNetwork(5), 'assets/prodAppConfig.json','assets/functions.js'),
-      // flavorInfo: Release(PrioritizeCache(),
-      //     'assets/prodAppConfig.json', 'assets/functions.js'),
-      environment: Environment.development.name,
-      // environmentInfo: Production(PrioritizeLocal(), 'assets/prodAppConfig.json','assets/functions.js'),
-      navigatorKey: null,
-      // developerConfig:
-      //     DeveloperConfig(enableChucker: false, proxyUrl: "192.168.119.217:9090"),
-      networkConfiguration:
-          NetworkConfiguration(defaultHeaders: {}, timeout: 30),
-      analytics: MyAnalytics()));
+  runApp(const DigiaUIExample());
 }
 
-class MyAnalytics extends DUIAnalytics {
+class DigiaUIExample extends StatelessWidget {
+  const DigiaUIExample({super.key});
+
   @override
-  void onDataSourceError(String dataSourceType, String source, errorInfo) {
-    print({
-      'dataType': dataSourceType,
-      'source': source,
-      'metaData': errorInfo.message
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      title: 'Digia UI Example',
+      debugShowCheckedModeBanner: false,
+      home: SplashScreen(),
+    );
+  }
+}
+
+class SplashScreen extends StatefulWidget {
+  const SplashScreen({super.key});
+
+  @override
+  State<SplashScreen> createState() => _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen>
+    with TickerProviderStateMixin {
+  late AnimationController _logoController;
+  late Animation<double> _logoScale;
+  DigiaUI? _digiaUI;
+  int _countdown = 3;
+
+  @override
+  void initState() {
+    super.initState();
+    _setupAnimation();
+    _initializeDigiaUI();
+  }
+
+  void _setupAnimation() {
+    _logoController = AnimationController(
+      duration: const Duration(milliseconds: 1500),
+      vsync: this,
+    );
+    _logoScale = Tween<double>(begin: 0.5, end: 1.0).animate(
+      CurvedAnimation(parent: _logoController, curve: Curves.elasticOut),
+    );
+    _logoController.forward().then((_) {
+      _startCountdown();
     });
   }
 
-  @override
-  void onDataSourceSuccess(
-      String dataSourceType, String source, metaData, perfData) {
-    print({
-      'dataType': dataSourceType,
-      'source': source,
-      'metaData': metaData.toString(),
-      'perfData': perfData.toString()
+  void _startCountdown() {
+    Future.delayed(const Duration(seconds: 1), () {
+      if (mounted) {
+        setState(() {
+          _countdown--;
+        });
+        if (_countdown > 0) {
+          _startCountdown();
+        } else {
+          _navigateToMainApp();
+        }
+      }
     });
   }
 
+  Future<void> _initializeDigiaUI() async {
+    try {
+      final initConfig = DigiaUIOptions(
+        accessKey: "68930cce1963e358762b546b",
+        flavor: Flavor.debug(environment: Environment.production),
+        networkConfiguration: const NetworkConfiguration(
+          defaultHeaders: {'X-App-Type': 'SimpleApp'},
+          timeoutInMs: 25000,
+        ),
+      );
+      final digiaUI = await DigiaUI.initialize(initConfig);
+      _digiaUI = digiaUI;
+    } catch (error) {
+      debugPrint('Initialization failed: $error');
+    }
+  }
+
+  void _navigateToMainApp() {
+    if (_digiaUI != null) {
+      Navigator.of(context).pushReplacement(
+        MaterialPageRoute(
+          builder: (context) => MainAppScreen(digiaUI: _digiaUI!),
+        ),
+      );
+    }
+  }
+
   @override
-  void onEvent(List<AnalyticEvent> events) {}
+  void dispose() {
+    _logoController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: const Color(0xFF3F51B5),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            AnimatedBuilder(
+              animation: _logoScale,
+              builder: (context, child) {
+                return Transform.scale(
+                  scale: _logoScale.value,
+                  child: Container(
+                    width: 150,
+                    height: 150,
+                    decoration: BoxDecoration(
+                      color: Colors.white,
+                      shape: BoxShape.circle,
+                      boxShadow: [
+                        BoxShadow(
+                          color: Colors.black.withValues(alpha: 0.2),
+                          blurRadius: 20,
+                          spreadRadius: 5,
+                        ),
+                      ],
+                    ),
+                    child: Image.asset('assets/logo.png'),
+                  ),
+                );
+              },
+            ),
+            const SizedBox(height: 40),
+            Text(
+              'Redirecting you to Digia in $_countdown',
+              style: const TextStyle(
+                color: Colors.white,
+                fontSize: 18,
+                fontWeight: FontWeight.w500,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class MainAppScreen extends StatelessWidget {
+  final DigiaUI digiaUI;
+
+  const MainAppScreen({super.key, required this.digiaUI});
+
+  @override
+  Widget build(BuildContext context) {
+    return DigiaUIApp(
+      digiaUI: digiaUI,
+      builder: (context) => DUIFactory().createInitialPage(),
+    );
+  }
 }
