@@ -5,10 +5,12 @@ import 'package:dio/dio.dart';
 import 'package:http_parser/http_parser.dart';
 import 'package:mime_type/mime_type.dart';
 
-import '../../../digia_ui.dart';
+import '../../dui_dev_config.dart';
 import '../../framework/data_type/adapted_types/file.dart';
+import '../../init/digia_ui_manager.dart';
 import '../../network/api_request/api_request.dart';
 import '../../network/core/types.dart';
+import '../../utils/logger.dart';
 
 class ApiHandler {
   static final ApiHandler _instance = ApiHandler._();
@@ -25,8 +27,7 @@ class ApiHandler {
       StreamController<Object?>? progressStreamController,
       CancelToken? cancelToken}) async {
     final stopwatch = Stopwatch();
-    final envVariables =
-        DigiaUIClient.instance.config.getEnvironmentVariables();
+    final envVariables = DigiaUIManager().environmentVariables;
     final envArgs = envVariables.map((key, value) {
       final dvalue = envVariables[key]?.defaultValue;
       return MapEntry('env.$key', dvalue);
@@ -38,7 +39,7 @@ class ApiHandler {
     }
 
     String url = _hydrateTemplate(apiModel.url, finalArgs);
-    final DigiaUIHost? host = DigiaUIClient.instance.developerConfig?.host;
+    final DigiaUIHost? host = DigiaUIManager().host;
     if (host is DashboardHost &&
         host.resourceProxyUrl != null &&
         url.startsWith('http:')) {
@@ -52,7 +53,7 @@ class ApiHandler {
         : null;
     final bodyType = apiModel.bodyType;
 
-    final networkClient = DigiaUIClient.getNetworkClient();
+    final networkClient = DigiaUIManager().networkClient;
     Response<Object?> response;
 
     stopwatch.start();
@@ -72,7 +73,7 @@ class ApiHandler {
               'total': p1,
               'progress': p0 / p1 * 100,
             });
-            print('Progress: ${p0 / p1 * 100}');
+            Logger.log('Progress: ${p0 / p1 * 100}', tag: 'ApiHandler');
           },
         );
       } else {
@@ -86,26 +87,26 @@ class ApiHandler {
       }
       stopwatch.stop();
       stopwatch.reset();
-      DigiaUIClient.instance.duiAnalytics?.onDataSourceSuccess(
-          'api',
-          url,
-          {'body': preparedData},
-          {'responseTime': stopwatch.elapsedMilliseconds});
+      // DigiaUIClient.instance.duiAnalytics?.onDataSourceSuccess(
+      //     'api',
+      //     url,
+      //     {'body': preparedData},
+      //     {'responseTime': stopwatch.elapsedMilliseconds});
       return response;
     } on DioException catch (e) {
       progressStreamController?.sink.addError(e);
 
-      DigiaUIClient.instance.duiAnalytics?.onDataSourceError(
-          'api',
-          url,
-          ApiServerInfo(e.response?.data, e.requestOptions,
-              e.response?.statusCode, e.error, e.message));
+      // DigiaUIClient.instance.duiAnalytics?.onDataSourceError(
+      //     'api',
+      //     url,
+      //     ApiServerInfo(e.response?.data, e.requestOptions,
+      //         e.response?.statusCode, e.error, e.message));
       rethrow;
     } catch (e) {
       progressStreamController?.sink.addError(e);
 
-      DigiaUIClient.instance.duiAnalytics?.onDataSourceError(
-          'api', url, ApiServerInfo(null, null, -1, e, null));
+      // DigiaUIClient.instance.duiAnalytics?.onDataSourceError(
+      //     'api', url, ApiServerInfo(null, null, -1, e, null));
       rethrow;
     } finally {
       progressStreamController?.sink.close();
