@@ -5,8 +5,9 @@ import 'package:flutter/services.dart';
 import 'package:flutter_js/flutter_js.dart';
 
 import '../../../digia_ui.dart';
-import '../../Utils/download_operations.dart';
-import '../../Utils/file_operations.dart';
+import '../../utils/download_operations.dart';
+import '../../utils/file_operations.dart';
+import '../../utils/logger.dart';
 import './js_functions.dart';
 
 class MobileJsFunctions implements JSFunctions {
@@ -42,7 +43,7 @@ class MobileJsFunctions implements JSFunctions {
           return true;
       }
     } catch (e) {
-      print('file not found');
+      Logger.error('file not found', tag: 'MobileJsFunctions', error: e);
       return false;
     }
   }
@@ -53,12 +54,13 @@ class MobileJsFunctions implements JSFunctions {
     JsEvalResult jsEvalResult =
         runtime.evaluate('JSON.stringify($fnName($input))');
     if (jsEvalResult.isError) {
-      if (DigiaUIClient.instance.developerConfig?.host is DashboardHost ||
-          kDebugMode) {
-        print('--------------ERROR Running Function-----------');
-        print('functionName ---->    $fnName');
-        print('input ----------> $input');
-        print('error -------> ${jsEvalResult.stringResult}');
+      if (DigiaUIManager().host is DashboardHost || kDebugMode) {
+        Logger.error('--------------ERROR Running Function-----------',
+            tag: 'MobileJsFunctions');
+        Logger.log('functionName ---->    $fnName', tag: 'MobileJsFunctions');
+        Logger.log('input ----------> $input', tag: 'MobileJsFunctions');
+        Logger.error('error -------> ${jsEvalResult.stringResult}',
+            tag: 'MobileJsFunctions');
       }
       throw Exception(
           'Error running function $fnName \n ${jsEvalResult.stringResult}');
@@ -78,12 +80,13 @@ class MobileJsFunctions implements JSFunctions {
     JsEvalResult promiseResult = await runtime.handlePromise(jsEvalResult);
 
     if (promiseResult.isError) {
-      if (DigiaUIClient.instance.developerConfig?.host is DashboardHost ||
-          kDebugMode) {
-        print('--------------ERROR Running Function-----------');
-        print('functionName ---->    $fnName');
-        print('input ----------> $input');
-        print('error -------> ${promiseResult.stringResult}');
+      if (DigiaUIManager().host is DashboardHost || kDebugMode) {
+        Logger.error('--------------ERROR Running Function-----------',
+            tag: 'MobileJsFunctions');
+        Logger.log('functionName ---->    $fnName', tag: 'MobileJsFunctions');
+        Logger.log('input ----------> $input', tag: 'MobileJsFunctions');
+        Logger.error('error -------> ${promiseResult.stringResult}',
+            tag: 'MobileJsFunctions');
       }
       throw Exception(
           'Error running function $fnName \n ${promiseResult.stringResult}');
