@@ -2,7 +2,7 @@ import 'dart:convert';
 import 'package:digia_ui/src/config/exception.dart';
 import 'package:digia_ui/src/config/factory.dart';
 import 'package:digia_ui/src/config/source/base.dart';
-import 'package:digia_ui/src/environment.dart';
+import 'package:digia_ui/src/init/flavor.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
@@ -35,9 +35,13 @@ void main() {
         )).thenAnswer((_) async {});
   });
 
-  ConfigSource createReleaseStrategy(InitPriority priority) {
+  ConfigSource createReleaseStrategy(DSLInitStrategy priority) {
     return ConfigStrategyFactory.createStrategy(
-      Release(priority, 'appConfig.json', 'functions.js'),
+      Flavor.release(
+        initStrategy: priority,
+        appConfigPath: 'appConfig.json',
+        functionsPath: 'functions.js',
+      ),
       mockProvider,
     );
   }
@@ -77,7 +81,8 @@ void main() {
           .thenAnswer((_) async => true);
 
       // ACT
-      final strategy = createReleaseStrategy(PrioritizeNetwork(5));
+      final strategy =
+          createReleaseStrategy(NetworkFirstStrategy(timeoutInMs: 5000));
       final config = await strategy.getConfig();
 
       // ASSERT
@@ -125,7 +130,8 @@ void main() {
       });
 
       // ACT
-      final strategy = createReleaseStrategy(PrioritizeNetwork(1));
+      final strategy =
+          createReleaseStrategy(NetworkFirstStrategy(timeoutInMs: 1000));
       final config = await strategy.getConfig();
 
       // ASSERT
@@ -161,7 +167,8 @@ void main() {
           .thenAnswer((_) async => json.encode(validConfigData));
 
       // ACT
-      final strategy = createReleaseStrategy(PrioritizeNetwork(1));
+      final strategy =
+          createReleaseStrategy(NetworkFirstStrategy(timeoutInMs: 1000));
       final config = await strategy.getConfig();
 
       verify(() => mockFileOps.readString('appConfig.json')).called(1);
@@ -190,7 +197,8 @@ void main() {
           .thenAnswer((_) async => json.encode(validConfigData));
 
       // ACT
-      final strategy = createReleaseStrategy(PrioritizeNetwork(5));
+      final strategy =
+          createReleaseStrategy(NetworkFirstStrategy(timeoutInMs: 5000));
       final config = await strategy.getConfig();
 
       verify(() => mockProvider
