@@ -1,8 +1,8 @@
-import 'package:digia_ui/src/Utils/asset_bundle_operations.dart';
-import 'package:digia_ui/src/Utils/download_operations.dart';
-import 'package:digia_ui/src/Utils/file_operations.dart';
 import 'package:digia_ui/src/config/provider.dart';
 import 'package:digia_ui/src/config/source/base.dart';
+import 'package:digia_ui/src/utils/asset_bundle_operations.dart';
+import 'package:digia_ui/src/utils/download_operations.dart';
+import 'package:digia_ui/src/utils/file_operations.dart';
 import 'package:mocktail/mocktail.dart';
 
 class MockConfigProvider extends Mock implements ConfigProvider {}
